Add base classes for screen and screen model
Introduced BaseScreen and BaseScreenModel abstract classes in presentation layer. This change will serve as a base for all screens and their models in the project. The BaseScreen and BaseScreenModel classes include common functionalities like handling the rendering of screens, listening to effects, errors management, and coroutines scopes. This will make the development more consistent and efficient by reducing the code duplication.